error of id type fixed on subject id in backend API

+++ back/src/main/java/com/openclassrooms/api/models/Subject.java
@@ -1,12 +1,6 @@
 package com.openclassrooms.api.models;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
+import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
@@ -15,17 +9,18 @@ public class Subject {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id; // Correction: passer à Long
+
     private String name;
     private String description;
     private LocalDateTime createdAt;
 
-    // Getters and Setters
-    public Integer getId() {
+    // Getters et Setters
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
